tests: move mock data to models directory.
