+++ test/unit/models/labs/certifications.mock.ts
@@ -1,0 +1,12 @@
+export const certificationDataMock = {
+  id: "ID",
+  ownerId: "OWNER_ID",
+  info: {
+    title: "TITLE",
+    issuer: "ISSUER",
+    month: "MONTH",
+    year: "YEAR",
+    description: "DESCRIPTION",
+    supportingDocument: "SUPPORTING_DOCUMENT",
+  }
+}
