+++ test/unit/models/labs/labs.mock.ts
@@ -1,0 +1,24 @@
+export const labDataMock = {
+  accountId: "ACCOUNT_ID",
+  services: [
+    "ID"
+  ],
+  certifications: [
+    "ID"
+  ],
+  verificationStatus: "Unverified",
+  info: {
+    boxPublicKey: "string",
+    name: "string",
+    email: "string",
+    phoneNumber: "string",
+    website: "string",
+    country: "string",
+    region: "string",
+    city: "string",
+    address: "string",
+    latitude: "string",
+    longitude: "string",
+    profileImage: "string",
+  }  
+}
