+++ test/unit/models/labs/orders.mock.ts
@@ -1,0 +1,25 @@
+export const orderDataMock = {
+  id: "ID",
+  serviceId: "SERVICE_ID",
+  customerId: "CUSTOMER_ID",
+  customerBoxPublicKey: "CUSTOMER_BOX_PUBLIC_KEY",
+  sellerId: "SELLER_ID",
+  dnaSampleTrackingId: "DNA_SAMPLE_ID",
+  currency: "DAI",
+  prices: [
+    {
+      component: "COMPONENT",
+      value: 0,
+    }
+  ],
+  additionalPrices: [
+    {
+      component: "COMPONENT",
+      value: 0,
+    }
+  ],
+  status: "Unpaid",
+  orderFlow: "RequestTest",
+  createdAt: 0,
+  updatedAt: 0
+}
