+++ test/unit/models/labs/services.mock.ts
@@ -1,0 +1,22 @@
+export const serviceDataMock = {
+  id: "ID",
+  ownerId: "OWNER_ID",
+  serviceFlow: "RequestTest",
+  info: {
+    pricesByCurrency: [
+      {
+        priceComponents: [
+          {
+            value: "VALUE",
+          }
+        ],
+        additionalPrices: [
+          {
+            value: "VALUE",
+          }
+        ],
+        currency: "CURRENCY"
+      }
+    ]
+  },
+}
