+++ test/unit/query/labs/certifications/certifications.query.mock.ts
@@ -5,16 +5,3 @@ export const certifications = {
     return returnToHumanMockWithParam(param);
   },
 }
-
-export const certificationDataMock = {
-  id: "ID",
-  ownerId: "OWNER_ID",
-  info: {
-    title: "TITLE",
-    issuer: "ISSUER",
-    month: "MONTH",
-    year: "YEAR",
-    description: "DESCRIPTION",
-    supportingDocument: "SUPPORTING_DOCUMENT",
-  }
-}
