+++ test/unit/query/labs/labs.query.mock.ts
@@ -5,28 +5,3 @@ export const labs = {
     return returnToHumanMockWithParam(param);
   },
 }
-
-export const labDataMock = {
-  accountId: "ACCOUNT_ID",
-  services: [
-    "ID"
-  ],
-  certifications: [
-    "ID"
-  ],
-  verificationStatus: "Unverified",
-  info: {
-    boxPublicKey: "string",
-    name: "string",
-    email: "string",
-    phoneNumber: "string",
-    website: "string",
-    country: "string",
-    region: "string",
-    city: "string",
-    address: "string",
-    latitude: "string",
-    longitude: "string",
-    profileImage: "string",
-  }  
-}
