+++ test/unit/query/labs/orders/orders.query.mock.ts
@@ -9,29 +9,3 @@ export const orders = {
     return returnToHumanMockWithParam(param);
   }
 }
-
-export const orderDataMock = {
-  id: "ID",
-  serviceId: "SERVICE_ID",
-  customerId: "CUSTOMER_ID",
-  customerBoxPublicKey: "CUSTOMER_BOX_PUBLIC_KEY",
-  sellerId: "SELLER_ID",
-  dnaSampleTrackingId: "DNA_SAMPLE_ID",
-  currency: "DAI",
-  prices: [
-    {
-      component: "COMPONENT",
-      value: 0,
-    }
-  ],
-  additionalPrices: [
-    {
-      component: "COMPONENT",
-      value: 0,
-    }
-  ],
-  status: "Unpaid",
-  orderFlow: "RequestTest",
-  createdAt: 0,
-  updatedAt: 0
-}
