+++ test/unit/query/labs/services/services.query.mock.ts
@@ -11,26 +11,3 @@ export const services = {
     return toHumanMock;
   }
 }
-
-export const serviceDataMock = {
-  id: "ID",
-  ownerId: "OWNER_ID",
-  serviceFlow: "RequestTest",
-  info: {
-    pricesByCurrency: [
-      {
-        priceComponents: [
-          {
-            value: "VALUE",
-          }
-        ],
-        additionalPrices: [
-          {
-            value: "VALUE",
-          }
-        ],
-        currency: "CURRENCY"
-      }
-    ]
-  },
-}
